Refactoring: Be more restrictive about transaction handling

+++ src/main/java/de/keybird/beagle/rest/ImportRestController.java
@@ -36,10 +36,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 import de.keybird.beagle.api.Document;
 import de.keybird.beagle.api.DocumentSource;
-import de.keybird.beagle.api.sources.ByteDocumentSource;
 import de.keybird.beagle.jobs.JobExecutionFactory;
 import de.keybird.beagle.jobs.JobExecutionManager;
-import de.keybird.beagle.jobs.persistence.DetectJobEntity;
+import de.keybird.beagle.jobs.source.ByteDocumentSource;
+import de.keybird.beagle.jobs.DetectJob;
 import de.keybird.beagle.repository.DocumentRepository;
 import de.keybird.beagle.rest.model.DocumentDTO;
 
@@ -73,13 +73,17 @@ public class ImportRestController {
     @RequestMapping(method = RequestMethod.POST, consumes = MediaType.APPLICATION_OCTET_STREAM_VALUE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void createDocument(@RequestParam(value = "name") String filename, HttpEntity<byte[]> requestEntity) {
-        final DocumentSource source = new ByteDocumentSource(filename, requestEntity.getBody());
-        jobExecutionManager.submit(new DetectJobEntity(source));
+        final DocumentSource source = new ByteDocumentSource(() -> filename, () -> requestEntity.getBody());
+        jobExecutionManager.submit(new DetectJob(source));
     }
 
     @RequestMapping(method = RequestMethod.DELETE)
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteAll() {
+        // TODO MVR this is ugly, but for now we kee it
+        if (jobExecutionManager.hasRunningJobs()) {
+            throw new IllegalStateException("Cannot delete documents while jobs are running");
+        }
         documentRepository.deleteAll();
     }
 
@@ -88,6 +92,10 @@ public class ImportRestController {
         final Document document = documentRepository.findOne(documentId);
         if (document == null) {
             return ResponseEntity.notFound().build();
+        }
+        // TODO MVR this is ugly, but for now we kee it
+        if (jobExecutionManager.hasRunningJobs()) {
+            throw new IllegalStateException("Cannot delete documents while jobs are running");
         }
         documentRepository.delete(document);
         return ResponseUtils.noContent();
